+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [2.0.2] / 2025-10-31
+### Features
+- Support `DataBuilder` in `RevitAddInsBuilder`. (Fix: #21)
+### Tests
+- Add `RevitAddInsBuilder_ManifestSettings_Tests`.
+- Remove `net6.0` target framework in tests.
+- Add `RevitAddInsBuilderManifestSettingsExtensions` with extension to create `ManifestSettings`.
+
 ## [2.0.1] / 2025-07-15
 ### Features
 - Support `Navisworks` extensible data.
@@ -100,6 +108,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - First Release
 
 [vNext]: ../../compare/1.0.0...HEAD
+[2.0.2]: ../../compare/2.0.1...2.0.2
 [2.0.1]: ../../compare/2.0.0...2.0.1
 [2.0.0]: ../../compare/1.0.6...1.2.0
 [1.0.6]: ../../compare/1.0.5...1.0.6